+++ src/lib.rs
@@ -1,7 +1,7 @@
 #![deny(clippy::all)]
 #![deny(missing_docs)]
 #![forbid(unsafe_code)]
-#![doc(html_root_url = "https://docs.rs/libunftp/0.18.7")]
+#![doc(html_root_url = "https://docs.rs/libunftp/0.18.8")]
 
 //! libunftp is an extensible, async, cloud orientated FTP(S) server library.
 //!
@@ -20,7 +20,7 @@
 //!
 //! ```toml
 //! [dependencies]
-//! libunftp = "0.18.7"
+//! libunftp = "0.18.8"
 //! unftp-sbe-fs = "0.2.0"
 //! tokio = { version = "1", features = ["full"] }
 //! ```